fixed most severe bugs

+++ CodenameProjectServer/RessourceKeeper.cs
@@ -8,18 +8,18 @@ namespace CodenameProjectServer
 {
     class RessourceKeeper
     {
-		public Int32 Wood {get; set;}
-		public Int32 Stone {get; set;}
+		public float Wood {get; set;}
+		public float Stone {get; set;}
 		public byte Faction {get; set;}
 
-        public RessourceKeeper(byte _faction, int _stones, int _wood)
+        public RessourceKeeper(byte _faction, float _stones, float _wood)
         {
             Faction = _faction;
             Stone = _stones;
             Wood = _wood;
         }
 
-        public void subtract(int stone, int wood){
+        public void subtract(float stone, float wood){
             Stone -= stone;
             Wood -= wood;
         }
